pyproject.toml
[tool.poetry]
name = "sora"
version = "0.1.0"
description = "Extendable GTK3 widgets"
authors = ["David Borzek <mail@davidborzek.de>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "sora" }]

[tool.poetry.scripts]
sora = "sora.cmd.main:main"

[tool.poetry.dependencies]
python = "^3.11"
libsass = "^0.22.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
PyGObject = "^3.46.0"
PyGObject-stubs = "^2.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
